Refactor tasklib/django.py to class based
The functions in django.py have now become application manager classes,
with a class variable called `tasks` that defines the list of tasks that
can be called from the command line (used in tasks.py).

As part of this, a lot of stuff that was stuffed in the global
environment variable becomes object variables.

Still todo:

- test it still works!
- put setup paths stuff in manager
- and allow debug log to go to file and different set of things to console
- use python logging for stuff
- get rid of `env[]` altogether (mostly just used for quiet/verbose, so logging
  should sort that out)