Replace most print statements with logging
This almost means we don't need to use env['quiet'] or env['verbose']
any more, though there are still a few bits to track down and eliminate.

I've used logging.warning for the short output stuff - should probably
do a more interesting logging thing, with that stuff having minimal
formatting ...

And add a --log-to-file option to log everything to file, whether the
console output is quiet, verbose or neither.